scripts/requirements.txt
elevenlabs>=1.0.0
google-genai>=1.0.0
